Chapitre 2 : corrections de Promise
